Refactor more utility functions
