+++ Assets/Scripts/Army/ArmyUtils.cs
@@ -4,6 +4,7 @@ using UnityEngine;
 using Mirror;
 using System.Collections.ObjectModel;
 using System.Linq;
+using Unity.Mathematics;
 
 public class ArmyUtils : MonoBehaviour
 {
@@ -79,6 +80,101 @@ public class ArmyUtils : MonoBehaviour
         List<Vector2> armyComplex;
         armyComplex = armyUnits.Select(i => i.GetIdentityZ()).ToList();
         return armyComplex;
+    }
+
+    /// <summary>
+    /// This function calculates the eigenvector as well as the centroid
+    /// </summary>
+    /// <param name="data"></param>
+    /// <param name="meanZ"></param>
+    /// <returns></returns>
+    public static (Vector2, Vector2) GetEigenCentroid(List<Vector2> data, Vector2 meanZ)
+    {
+        var xMean = meanZ.x;
+        var yMean = meanZ.y;
+
+        var varX = 0f;
+        varX = data.Aggregate(varX,
+            (current, c) =>
+            {
+                return current + (c.x - xMean) * (c.x - xMean);
+            }) / (data.Count - 1);
+
+        var varY = 0f;
+        varY = data.Aggregate(varY,
+            (current, c) =>
+            {
+                return current + (c.y - yMean) * (c.y - yMean);
+            }) / (data.Count - 1);
+
+        var covXY = 0f;
+        covXY = data.Aggregate(covXY,
+            (current, c) =>
+            {
+                return current + (c.y - yMean) * (c.x - xMean);
+            }) / (data.Count - 1);
+
+        // the cov matrix is [varX, covXY; covXY, varY]
+        // now we calculate the eigenvectors and eigenvalues
+
+        var delta = math.sqrt((varX + varY) * (varX + varY) - 4 * (varX * varY - covXY * covXY));
+        var l1 = (varX + varY + delta) / 2;
+        var l2 = Mathf.Abs(varX + varY - delta) / 2;
+
+        var l = Mathf.Max(l1, l2);
+        // the new thing to solve is then
+        // [varX - l, covXY; covXY, varY - l][a; b] = 0
+
+        var aFactor = varX - l + covXY;
+        var bFactor = covXY + varY - l;
+
+        var eigenVector = new Vector2(bFactor, -aFactor).normalized;
+        return (eigenVector, new Vector2(xMean, yMean));
+    }
+
+    /// <summary>
+    /// This function does a PCA on the units identities
+    /// and splits the army along the principle axis
+    /// the split is returned as a tuple of lists
+    /// </summary>
+    /// <param name="armyUnits"></param>
+    /// <param name="armyComplex"></param>
+    /// <param name="meanZ"></param>
+    /// <param name="splitThreshold"></param>
+    /// <returns>Two lists of Units representing the two splits</returns>
+    public static (List<Unit>, List<Unit>) CalculateSplit(List<Unit> armyUnits, List<Vector2> armyComplex, Vector2 meanZ, float splitThreshold)
+    {
+        // GetEigenCentroid returns (eigenvector, centroid) of the dataset
+        var (eigenVector, centroid) = ArmyUtils.GetEigenCentroid(armyComplex, meanZ);
+
+        // if we get a perfectly circular set it is possible for eigen to be 0
+        // just randomly pick an axis if so
+        if (eigenVector == Vector2.zero)
+        {
+            var angle = UnityEngine.Random.Range(0, Mathf.PI * 2);
+            eigenVector = new Vector2(Mathf.Cos(angle), Mathf.Sin(angle));
+        }
+
+        // project the centroid onto the eigen
+        var centroidProj = Vector2.Dot(centroid, eigenVector);
+
+        var retSplit1 = new List<Unit>();
+        var retSplit2 = new List<Unit>();
+
+        for (int i = 0; i < armyComplex.Count; i++)
+        {
+
+            if (Vector2.Dot(armyComplex[i], eigenVector) < centroidProj)
+            {
+                retSplit1.Add(armyUnits[i]);
+            }
+            else
+            {
+                retSplit2.Add(armyUnits[i]);
+            }
+        }
+
+        return (retSplit1, retSplit2);
     }
 
     // This is a function that maps a value from one range to another
